UI: Remove File Browser "small_size"
Remove non-working feature that changes File Browser list view when the
thumbnails are smaller than 64x64.

---

File Browser code currently contains a feature for "small_size" which doesn't work and causes some issues.

The intent is to alter the File Browser listing to show a more compact display under some conditions. Unfortunately it is tied to the size of thumbnails. Which makes no sense since this is for the non-thumbnail view. And the condition is when the thumbnail size drops below 64, which has never been the case so this code has never been used or tested.

But since #105815 we are allowing thumbnails to get smaller than 64, which exposes the problems with the "small_size". It alters the width of the date column but does not properly alter the string that goes in there. It alters the width of the size column but then we don't have a more compact form of size.  The result is that if you have thumbnails smaller than 64 then some columns in list view will overflow.

The PR just removes this feature, rather than fix it. A fix is submitted as #112464 but it is getting a bit late for that. We could merge this one and perhaps revive the feature for 4.1