Fix 111532: Texture Painting on Byte Data With Texture Limit
When painting on a byte texture with data colorspace and texture limit
enabled `gpu_texture_update_scaled` expects an input buffer clipped to
the bounds of the input image being updated. In case for Byte Data
textures it received the full byte buffer and incorrect data was
uploaded to the GPU.

This PR fixes this by not selecting the optimized code-branch when the
texture is a Byte buffer with Data color space and scaling should
happen.