Fix: Continuous asset reloading in edit mode with no node tools
The drawing code checked if the node tool asset tree was empty to decide
whether to rebuild the tree. Instead store a dirty state, so we can skip
rebuilding when no node tools were found.

Theoretically the asset index makes this operation not terribly slow,
but it's still bad to do it every redraw even so.