Vulkan: Use Correct Point Shaders
In Vulkan (and Metal) it is not possible to use a constant as point
size. It needs to be set for each vertex when drawing points. Blender
has specialized shaders for that, but not all code respect those
shaders.

This PR will add an assert inside the vulkan backend when incorrect
usage of shaders are detected.