USD: Add MaterialX shader export
This change adds the ability to export MaterialX networks into the resulting USD layer.

Details:

A new export option has been added to the USD export to enable MaterialX export. It is off by default currently due to reasons in the caveats section.

When enabled, it exports the MaterialX shading network alongside the UsdPreviewSurface network, on the same USD Material. This allows the same material to be used by renderers that don't support MaterialX, using the USDPreviewSurface as a fallback. This is similar to setups in other DCC packages, and matches the format we've used in our Reality Composer Pro asset library.

It uses the existing MaterialX framework used to generate MaterialX documents for rendering, to act as the basis for the USD graph. In this process it also re-uses the existing texture export code as well if provided and necessary.

Once the MaterialX document is created, I use usdMtlx to generate a USD shading network. Unfortunately, usdMtlx generates a graph that is unlike what other DCCs that support MaterialX-embedded-in-USD generates. It generates several extra prim hierarchies, and externalizes all shader inputs, making them difficult to edit in other MaterialX graph editors.

To workaround this, I generate the MaterialX shading network onto a temporary stage, where I then run various pre-processing steps to prevent prim collisions and to reflow the paths once they're converted.

The PrimSpecs are then copied over to their new path. The resulting prim hierarchy matches what many artists we've worked with prefer to work with.

Caveats:

I have opted to leave the Export MaterialX checkbox off by default. When using the Principled Shader, the resulting graph is very usable.
However, when using some of the other BSDFs, the shading networks generated by the existing MaterialX framework in Blender generate some shading graphs that are difficult for usdview and other DCC's to understand. The graph is still correct, but because I am trying to prioritize compatibility, I feel the default should be off.

In future PRs we can aim to make the graphs for those other BSDFs play better with other DCCs.

Other Implementation Details:

As part of this commit we've also done the following:

* We place some of the materialx graphs inside a passthrough nodegraph to avoid node conflicts
* We better handle some shader output types , and better handle some conflict cases
* I've moved the ExportTextureFunction to materials.h due to some difficult to resolve header ordering issues. This has no effect on any runtime code.
* There is a test for the MaterialX export that does some basic checking to make sure we get an export out the other end that matches our expectations

Authored by Apple: Dhruv Govil

This PR is based on an earlier implementation by Brecht van Lommel , as well as Brian Savery and his teams' work at AMD to implement the general MaterialX framework within Blender.