Fix panel drawing when using multiple modifiers.
The child panels were using the same UI block for each instance of the
modifier, because the identifiers were not unique.

Passing the parent panel optionally allows generating a more unique
string for blocks. Note that this will still fail when nesting more than
two levels of instanced child panels, but that's not a case which exists
at this time. A future fix can address this more thoroughly.