Remove empty space in tool settings header with toolbar visible
So far, when the toolbar was visible, the tool settings header would
show empty space instead of the tool icon. This was probably to avoid
jumping of the layout. However this empty space looks quite jarring, and
its size doesn't even match the default toolbar width, so alignment
would look off (causing visual noise). On the other hand the jumping
seems like a minor problem, it's quite suttle, esp since the viewport
overlay text does a similar jump. Having the icon appear and disappear
based on the toolbar visibility also causes visual distraction, so the
problem is was mitigated at best.

Alternatively, we could always display the tool icon, but that means the
same icon is displayed twice. By default the first tool is selected, so
you have twice the same icon in close proximity.