Fix: Missing node tool menu update after moving asset to catalog
Moving an asset from "Unassigned" to a catalog needs to clear the node tool
asset tree caches and redraw the editor. That's solved here by emitting
an asset-specific notifier rather than the less clear "file list" one.

---

I'm starting to think that it would be better to have a more direct
connection betweent the asset system and the asset tree caches.
Maybe it would make sense to register them with the asset system
which could then clear them at any time. That would resolve this
mess with notifiers, since the asset system should know everything
about when assets change. But I'm not quite sure which part of the
asset system would do that.