Fix: Workbench Reads/Writes to Uninitialized Texture
TAA in Workbench writes back to the depth in front texture when needed.
The depth in front texture passed via the parameters however can be
nullptr. In Workbench there are 2 depth in front textures, One from the
GPU Viewport, and one stored in the resources. The one in the resources
are always allocated. The one in the viewport might not.

It is unclear to me if the viewport texture in this case is intended to
be allocated, or what this PR does ignored when copy the results back to
the viewport.

`GPU_texture_copy` assumes that both the source and destination texture
are valid textures.

This was detected as the Vulkan backend is less forgiven when passing
incorrect parameters as it is still under development.