I18n: add label declaration to node sockets so they can be shortened
In !112591, nodes got the ability to group sockets into panels. The
labels for these sockets are automatically shortened if they begin
with the same text as their parent labels. For instance, "Transmission
Weight" will be shortened to just "Weight" because it is under the
"Transmission" panel.

While this is a good heuristic for English, it breaks down in
languages which do not have the same word order.

This commit adds a `.label()` callback to socket declarations so that
a shortened label can be explicitly declared. It reuses the `label`
field in `bNodeSocket`, which is currently used to assign dynamic
labels at runtime in some nodes.

It also adds two regexps to the translation script so that these new
fields can be extracted to the .po translation files. One extracts the
label with a translation context, the other without.
Only the one without context is currently in use.

Fixes #112970: Node socket labels under panels are not shortened when
translated.