UI: Hide asset shelf toggles if not available
The "Asset Shelf" checkbox can be confusing for people, they enable it
but it still can't get the shelf to show up. This is because it is only
available in certain context (pose mode only for the start). For now
hide the option entirely unless the asset shelf becomes available.

This change adds a new `bpy.types.Region.poll_failed` property to query
the availability of a region.