Fix #113038: Box projection rotation inconsistent for dominant Z axis
While rotation might seem a bit arbitrary in box, it seems it would make
sense to match orientation we would get from flat projection.

This is want this PR does (for Cycles [incl OSL] and EEVEE).

As this might change the look of many old files, the motivation should
really have its justification (not totally sure about this, but previous
behavior does seem inconsistent and 4.0 might be the chance to rectify).