Node Editor: Add overlay for reroute auto labels
Add an overlay option to automatically display a label on reroute nodes.
This automatic label is propagated through chained reroute nodes and
is based on the socket the reroutes are linked to. If a reroute node is
explicitly labeled, that label is propagated instead.

The automatic label is dimmed to distinguish it from manual labels.