Anim: Align the bone of 'Shift-A > Armature' with the Y-axis
Align the default bone that's added to the Armature you get with
'Shift-A > Armature' to the Y-axis. When the alignment is set to 'World'
this will actually align the bone's axes with the world axes.

Note that the 'View' alignment option already aligned the bone along the
Y-axis, so the code has been simplified a bit.