Anim: align bone with world Y in Armature Edit mode
Pressing Shift+A (Add > Single Bone) in Armature Edit mode will now
align the bone with the world Y-axis instead of the Z-axis. This makes
the bone's axes align with the world axes (instead of aligning bone-Y
with world-Z).

This only applies when using 'Align To: World' in the Editing
preferences. 'Align To: View' still behaves as before, and 'Align To: 3D
Cursor' behaves like 'World' because that option was never handled
anyway.