Mesh: Add flag to store presence of overlapping topology
For improving performance in the triangulate node, it's very helpful to
know whether the mesh has edges or faces that overlap each other.
If that is known to be false, the node can skip edge deduplication.

Since this property doesn't often change as part of mesh processing
operations, it's simple to store it once. Currently this is done using
a `SharedCache`, though there is no lazy computation since I'm not
sure it would be worth it.