I18n: translate Move to Collection operator menu
Several items are untranslated in the Move to Collection operator
menu (the armature version, `ARMATURE_OT_move_to_collection`). This is
because it uses custom drawing to allow the menu to be dynamically
generated.

Because of this custom drawing, some labels have to be translated
manually using `CTX_IFACE_()`. The translation context is "Operator"
because those messages come from an operator.