Nodes: Use index instead of reordering for draw order
Currently nodes are reordered so that the "on top" nodes are last in the
list. Node order changing for simple operations like selection means we
either have to reevaluate the node tree data-block on selections or
accept that the evaluated order can be different from the original.
Currently we do the latter (see d76a0e98baecfca4e691), but makes it
complex to refer to nodes by index, and is hard to reason about.

Instead of reordering nodes, store the ui order in the node itself
and sort the nodes before drawing them or doing any processing
that depends on the "depth."