Select Engine: Support object occlusion when selecting in edit mode
This fixes #112739

Remove the condition to use only on retopology overlay.

---
The code already exists for retopology, so there isn't much to review.

This PR is more for a quick discussion whether it is worth changing this old behavior.

This is a recurring issue on the tracker.
Many (if not all) users working on the Donut tutorial seem to face this issue of non-visible geometry being selected.

Therefore, it seems important to resolve this situation in some way, whether by displaying this geometry or supporting occlusion.

This PR implements the occlusion solution.