Cycles: Add option to control smoothing when using bump map
Cycles implements the "Taming the Shadow Terminator" paper by
Matt Jen-Yuan Chiang to solve shadow terminator issues when a
bump map is applied.

This correction term is applied unconditionally, which makes
it harder to have full control over shading via normals for
stylistic reasons.

This change exposes this corrective term as an option called
"Smooth Bump Illumination" which is available in the shader
settings next to the "Transparent Shadows".

The reason to make it per-shader rather than per-object is to
allow flexibility of a control: it is possible that an object
has multiple shaders attached to it, and only some of them
used for bump mapping.

The option is enabled by default, so there is no changes for
existing setups.